Substitue Dr.Bharrat with Bharrat
This is a special case where the doctor's name is not
separated from their title.

The member's URL displays their name as: 'Dr.Bharrat Jagdeo'

Since the title and name are not separated by a space, as is
the case with all other members, the scrape cannot yet handle them
properly.

As is the case with the scraper before refactoring, we are simply
substituting the Dr. part out from the name.